added support for the extent_values_filter property to be added to categories/layers in the map-vis-layers.json document.  This allows specific layers or categories to filter acceptable extent values as reported by the geoserver to subsets of those values for specific cases.
For example some current year layers allow for selection of "forecast" dates in the future
but we'd like to restrict the user from selecting future dates and so can filter extent values
to those prior to "today"

Or for some spring index layers where the dates of interest are before a given date (e.g. May 02)
while the geo server reports support for all days of the year (even though the layer doesn't visually
change after a given date so it makes no real logical sense to allow the user to select those dates).

Also updated the ngdocs to capture these details a bit more so that in the future hopefully supporting
or extenting this functionality requires less reverse engineering.