Remove the use of extent_values_filter for many forecast map layers.
For many layers the previous extent_values_filter was configured like:
"extent_values_filter": {
            "name": "extentDates",
            "args": [null,"today"]
        },
Which would exclude forecast dates (those beyond "today") from what the
server is willing to supply data for.

Notes:

The si-x:leaf/bloom_anomaly maps did NOT have this filter in place so were
always allowing for view of "forecast" data

The feature notes are very explicit about how many days into the future to
allow display for but this detail (now that the filter is removed) is
entirely in the hands of the server.

Today is April 3 and MOST (but not all) of the layers will now show map
data up to April 9.  Some layers however only expose data up to April 8.

Lastly the "default" value for which day to show is similarly under the control
of the server (it returns which date should be the default) and the server is
returning the latest date.  I.e. while previously these layers would show
"today" as the default they will now show the default as returned by the server
(for today, April 3, either April 8 or 9 depending on the layer selected).