Initial stab at only yes data
Due to the ability to toggle yes/no for each species/network things can
get a bit confusing when intersecting the two.  Counts are indepentent.

Example:
1.) leave the global setting as the default (all data)
2.) filter "Red Maple" with "A.T. Seasons" for 2016-2017 (red maple intersects with what A.T. Seasons gathers.

Now individually toggle the "Yes Data Only" checkboxes found within the species/network filter arguments.

The top level numbers are driven by "species" so the species arg/date arg will have the same number.
When filtering the species by yes only then the network # will decrease by some amount due to stations that have no yes records.
This isn't obvious.

If you do the reverse, "yes only" for the station, while "all" for the species, then only the counts on the station will decrease.

We cannot attach the "yes only" for the station to a species since this would potentially break filters with multiple networks that
intersect on a given species.

In the above "simple" case the functionality works relatively logical but is not intuitive when the settings accross arguments aren't
in alignment.