Start on AGDD range slider
Notes:
As part of this work the base slider implementation needed to be
replaced since the one in use for over-all opacity wasn't robust enough
for this use case WRT labeling.  The new slider implementation likely
will require some CSS work since the slider handle is large.

At this moment the control is available for any map layer, not just the two
specified in the requirements.

The requirements document states that the lower opacity should be 20%.  This seems to
result in a visible "box" around the lower 48.  This lower opacity is currently 0.